Added Database interface and Updated code sharing strategy for SQLiteAdapter
