+++ packages/datastore-storage-adapter/src/ExpoSQLiteAdapter/ExpoSQLiteDatabase.ts
@@ -2,6 +2,7 @@ import * as SQLite from 'expo-sqlite';
 import { ConsoleLogger as Logger } from '@aws-amplify/core';
 import { PersistentModel } from '@aws-amplify/datastore';
 import { ParameterizedStatement } from '../commons/SQLiteUtils';
+import { Database } from '../commons/Database';
 
 const logger = new Logger('ExpoSQLiteDatabase');
 
@@ -31,24 +32,32 @@ get the result of an `executeSql` command inside of a transaction
 
 */
 
-class ExpoSQLiteDatabase {
+class ExpoSQLiteDatabase implements Database {
 	private db: SQLite.WebSQLDatabase;
 
 	public async init(): Promise<void> {
-		this.db = SQLite.openDatabase(DB_NAME, DB_VERSION, DB_DISPLAYNAME, DB_SIZE);
+		//only openDatabase once.
+		if (!this.db) {
+			this.db = SQLite.openDatabase(
+				DB_NAME,
+				DB_VERSION,
+				DB_DISPLAYNAME,
+				DB_SIZE
+			);
+		}
 	}
 
-	public async createSchema(statements: string[]) {
+	public async createSchema(statements: string[]): Promise<void> {
 		return await this.executeStatements(statements);
 	}
 
-	public async clear() {
+	public async clear(): Promise<void> {
 		logger.debug('Deleting database');
 		//delete database is not supported by expo-sqlite. alternative way is to get all table names and drop them.
 		//await SQLite.deleteDatabase(DB_NAME);
 		this.dropAllTables();
 		logger.debug('Database deleted');
-		await this.closeDB();
+		//await this.closeDB();
 	}
 
 	public async get<T extends PersistentModel>(
@@ -70,7 +79,6 @@ class ExpoSQLiteDatabase {
 				});
 			}
 		);
-
 		const result =
 			resultSet &&
 			resultSet.rows &&
@@ -101,13 +109,11 @@ class ExpoSQLiteDatabase {
 				});
 			}
 		);
-
 		const result =
 			resultSet &&
 			resultSet.rows &&
 			resultSet.rows.length &&
 			resultSet.rows._array;
-
 		return result || [];
 	}
 
@@ -117,10 +123,9 @@ class ExpoSQLiteDatabase {
 				tx.executeSql(
 					statement,
 					params,
-					() => resolve(true),
+					() => resolve(),
 					(_, error) => {
 						reject(error);
-
 						return true;
 					}
 				);
@@ -128,7 +133,9 @@ class ExpoSQLiteDatabase {
 		});
 	}
 
-	public async batchQuery(queryStatements: Set<ParameterizedStatement>) {
+	public async batchQuery(
+		queryStatements: Set<ParameterizedStatement>
+	): Promise<any[]> {
 		const results = [];
 		await new Promise((resolve, reject) => {
 			this.db.readTransaction(function (tx) {
@@ -141,22 +148,20 @@ class ExpoSQLiteDatabase {
 						},
 						(_, error) => {
 							reject(error);
-
 							return true;
 						}
 					);
-					resolve(true);
 				}
+				resolve();
 			});
 		});
-
 		return results;
 	}
 
 	public async batchSave(
 		saveStatements: Set<ParameterizedStatement>,
 		deleteStatements?: Set<ParameterizedStatement>
-	) {
+	): Promise<void> {
 		await new Promise((resolve, reject) => {
 			this.db.transaction(function (tx) {
 				for (const [statement, params] of saveStatements) {
@@ -170,6 +175,7 @@ class ExpoSQLiteDatabase {
 						}
 					);
 				}
+
 				if (deleteStatements) {
 					for (const [statement, params] of deleteStatements) {
 						tx.executeSql(
@@ -183,7 +189,7 @@ class ExpoSQLiteDatabase {
 						);
 					}
 				}
-				resolve(true);
+				resolve();
 			});
 		});
 	}
@@ -191,7 +197,7 @@ class ExpoSQLiteDatabase {
 	public async selectAndDelete(
 		query: ParameterizedStatement,
 		_delete: ParameterizedStatement
-	) {
+	): Promise<any[]> {
 		let results = [];
 
 		const [queryStatement, queryParams] = query;
@@ -199,11 +205,12 @@ class ExpoSQLiteDatabase {
 
 		results = await new Promise((resolve, reject) => {
 			this.db.transaction(function (tx) {
+				var tempresults;
 				tx.executeSql(
 					queryStatement,
 					queryParams,
 					(_tx, res) => {
-						resolve(res.rows._array);
+						tempresults = res.rows._array;
 					},
 					(_, error) => {
 						reject(error);
@@ -219,14 +226,14 @@ class ExpoSQLiteDatabase {
 						return true;
 					}
 				);
+				resolve(tempresults);
 			});
 		});
-
-		return results;
+		return results || [];
 	}
 
 	private async executeStatements(statements: string[]): Promise<void> {
-		return await new Promise((resolve, reject) => {
+		await new Promise((resolve, reject) => {
 			this.db.transaction(function (tx) {
 				for (const statement of statements) {
 					tx.executeSql(
@@ -239,7 +246,6 @@ class ExpoSQLiteDatabase {
 						}
 					);
 				}
-
 				resolve();
 			});
 		});
@@ -248,19 +254,18 @@ class ExpoSQLiteDatabase {
 	private async closeDB() {
 		if (this.db) {
 			logger.debug('Closing Database');
-
-			//closing database is not supported by expo-sqlite
-			//await this.db.close();
+			//closing database is not supported by expo-sqlite. Workaround is to access the private db variable and call the close() method.
+			var tempdb: any = this.db;
+			await tempdb._db.close();
 			logger.debug('Database closed');
 		}
 	}
 
-	private async dropAllTables() {
-		//to be implemented. Returns array of table names.
-		const tableNames: string[] = await new Promise((resolve, reject) => {
+	private async getAllTableNames(): Promise<string[]> {
+		const resultSet: any[] = await new Promise((resolve, reject) => {
 			this.db.transaction(tx => {
 				tx.executeSql(
-					'SHOW TABLES',
+					`SELECT name FROM sqlite_master where type='table'`,
 					[],
 					(tx, results) => {
 						resolve(results.rows._array);
@@ -273,9 +278,16 @@ class ExpoSQLiteDatabase {
 			});
 		});
 
+		const tableNames = resultSet.map(e => e.name);
+		return tableNames;
+	}
+
+	private async dropAllTables() {
+		const tableNames = await this.getAllTableNames();
+
 		for (const tableName of tableNames) {
 			const statement = `DROP TABLE IF EXISTS ${tableName}`;
-			this.executeStatements([statement]);
+			await this.executeStatements([statement]);
 		}
 	}
 }
