+++ packages/datastore-storage-adapter/src/index.ts
@@ -1,2 +1,3 @@
 import ExpoSQLiteAdapter from './ExpoSQLiteAdapter/ExpoSQLiteAdapter';
-export { ExpoSQLiteAdapter };
+import SQLiteAdapter from './SQLiteAdapter/SQLiteAdapter';
+export { SQLiteAdapter, ExpoSQLiteAdapter };
